--- a/line_benchmark_test.go
+++ b/line_benchmark_test.go
@@ -47,7 +47,6 @@ func benchmarkLineToEvents(times int, b *testing.B) {
 	}
 }
 
-
 func BenchmarkLineToEvents1(b *testing.B) {
 	benchmarkLineToEvents(1, b)
 }