Adds parsing for SignalFX-style tagsand a benchmark for LineToEvents. Fixes #268. It's unclear from the signalfx docs, but it appears that the tags can exist anywhere in the name field, making parsing a little more complicated.

There is a definite slight cost to parsing these tags since they can be in the middle of the name, and have opening and closing runes.

Before:
```
go test -bench='BenchmarkLineToEvents(1|5|50)'                                                                                                                
goos: darwin
goarch: amd64
pkg: github.com/prometheus/statsd_exporter
BenchmarkLineToEvents1-12         148065              7847 ns/op
BenchmarkLineToEvents5-12          30534             39436 ns/op
BenchmarkLineToEvents50-12          3038            392718 ns/op
PASS
ok      github.com/prometheus/statsd_exporter   4.261s

```

After:
```
go test -bench='BenchmarkLineToEvents(1|5|50)'                                                                                                                 
goos: darwin
goarch: amd64
pkg: github.com/prometheus/statsd_exporter
BenchmarkLineToEvents1-12         132652              8131 ns/op
BenchmarkLineToEvents5-12          29119             40593 ns/op
BenchmarkLineToEvents50-12          2988            405923 ns/op
PASS
ok      github.com/prometheus/statsd_exporter   4.357s

```