@glightfoot 2020-06-16
Note: if we move all the metrics into their own package, we can have the benchmark tests live in the package dirs instead at the top-level. Happy to do that too if necessary