@glightfoot 2020-06-17
As-is, this does not check for any other tag styles if it find SignalFx tags