@glightfoot 2020-06-19
Without knowing the acceptable characters, I thought the best way to handle it would be to pass the name on unchanged, but then the other tag parsing catches the ',' and splits it up trying to parse InfluxDB tags. If we want to short-circuit tag parsing if there's a non-matching bracket, I can do that, but do we want to drop the line, or just pass it through? In theory, you could have a valid InfluxDB tagged line with a bracket in the name... `foo[strange.name,09digits=0,tag.with.dots=1:100|c`