@matthiasr 2020-06-19
I think the best would be passing it through, with a debug log line + metric increment if it _almost_ looks like signalfx tags but isn't. that way, existing behavior is unchanged for anything that isn't a fully formed signalfx tag.

I would rather not have the exclusive switching, I'm sure there is someone out there using both the dogstatsd and the librato client …