@glightfoot 2020-06-19
Made the change - is this the behavior you were expecting?

```
[tag1=bar,tag2=bazfoo.test:100|c
 &event.CounterEvent{CMetricName:"[tag1=bar,tag2=bazfoo.test", CValue:100, CLabels:map[string]string{}}

tag1=bar,tag2=baz]foo.test:100|c
&event.CounterEvent{CMetricName:"tag1=bar,tag2=baz]foo.test", CValue:100, CLabels:map[string]string{}}

foo.tag1=bar,tag2=baz]test:100|c
&event.CounterEvent{CMetricName:"foo.tag1=bar,tag2=baz]test", CValue:100, CLabels:map[string]string{}}
```